change prune tech again to support archive properly
we now follow the mtime of the archive rather than link its age to log.
this allows us to clear archive/ even if log doesn't exist anymore.

this is how we should move forward on this as we might need to prune
other bits in the future.

the prune function looks well shit now

+++ lib/jenkins/jobdir.rb
@@ -41,16 +41,18 @@ module Jenkins
 
       # Filter now locked directories
       content.reject! { |d| locked.include?(File.realpath(d)) }
-      # Filter directories that have no log anymore
-      content.reject! { |d| !File.exist?("#{d}/log") }
 
       content.sort_by! { |c| File.basename(c).to_i }
       content[0..-6].each do |d| # Always keep the last 6 builds.
+        max_age = 14
         log = "#{d}/log"
         archive = "#{d}/archive"
-        next unless age(log) > 14
-        FileUtils.rm(File.realpath(log))
-        FileUtils.rm_rf(File.realpath(archive)) if File.exist?(archive)
+        if File.exist?(log)
+          FileUtils.rm(File.realpath(log)) if age(log) > max_age
+        end
+        if File.exist?(archive)
+          FileUtils.rm_r(File.realpath(archive)) if age(archive) > max_age
+        end
       end
     end
   end
